refactor: create internal sharable typescript config workspace
