+++ packages/core/package.json
@@ -4,6 +4,7 @@
   "description": "Accessible react component examples",
   "homepage": "https://docs.a11y-react.com",
   "license": "MIT",
+  "type": "module",
   "main": "lib/cjs/index.js",
   "module": "lib/esm/index.js",
   "sideEffects": false,
@@ -48,6 +49,7 @@
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
+    "@internal/typescript-config": "workspace:*",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@types/styled-components": "^5.1.26",
@@ -55,6 +57,7 @@
     "react-dom": "^18.3.1",
     "react-is": "^18.3.1",
     "styled-components": "^5.3.5",
+    "typescript": "^5.7.2",
     "vitest": "^2.1.8"
   }
 }
