+++ packages/ui/tsconfig.json
@@ -1,7 +1,12 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["src", "index.ts"],
-  "exclude": ["stories"],
+  "include": [
+    "src",
+    "index.ts"
+  ],
+  "exclude": [
+    "../../stories"
+  ],
   "compilerOptions": {
     "declaration": true,
     "rootDir": "src",
